Deepen match retrieval module
